third_party/WebKit/LayoutTests/http/tests/history/cross-origin-replace-history-object-child-expected.txt

ALERT: PASS: Access to window.frames[0].history threw an exception.
ALERT: About to shadow child window's history object: [object History]
ALERT: Shadowed child window's history object: 
ALERT: PASS: Access to window.frames[0].history threw an exception.
